+++ src/commands/hardis/project/deploy/sources/metadata.ts
@@ -1,64 +1,64 @@
 /* jscpd:ignore-start */
 
-import { flags, SfdxCommand } from "@salesforce/command";
-import { Messages } from "@salesforce/core";
-import { AnyJson } from "@salesforce/ts-types";
-import * as c from "chalk";
-import * as fs from "fs-extra";
-import * as os from "os";
-import * as path from "path";
-import { MetadataUtils } from "../../../../../common/metadata-utils";
-import { checkSfdxPlugin, execCommand, uxLog } from "../../../../../common/utils";
-import { CONSTANTS, getConfig } from "../../../../../config";
+import { flags, SfdxCommand } from '@salesforce/command';
+import { Messages } from '@salesforce/core';
+import { AnyJson } from '@salesforce/ts-types';
+import * as c from 'chalk';
+import * as fs from 'fs-extra';
+import * as os from 'os';
+import * as path from 'path';
+import { MetadataUtils } from '../../../../../common/metadata-utils';
+import { checkSfdxPlugin, execCommand, uxLog } from '../../../../../common/utils';
+import { CONSTANTS, getConfig } from '../../../../../config';
 
 // Initialize Messages with the current plugin directory
 Messages.importMessagesDirectory(__dirname);
 
 // Load the specific messages for this file. Messages from @salesforce/command, @salesforce/core,
 // or any library that is using the messages framework can also be loaded this way.
-const messages = Messages.loadMessages("sfdx-hardis", "org");
+const messages = Messages.loadMessages('sfdx-hardis', 'org');
 
 export default class DxSources extends SfdxCommand {
-  public static title = "Deploy metadata sources to org";
+  public static title = 'Deploy metadata sources to org';
 
-  public static description = messages.getMessage("deployMetadatas");
+  public static description = messages.getMessage('deployMetadatas');
 
-  public static examples = ["$ sfdx hardis:project:deploy:sources:metadata"];
+  public static examples = ['$ sfdx hardis:project:deploy:sources:metadata'];
 
   protected static flagsConfig = {
     check: flags.boolean({
-      char: "c",
+      char: 'c',
       default: false,
-      description: messages.getMessage("checkOnly")
+      description: messages.getMessage('checkOnly')
     }),
     packagexml: flags.string({
-      char: "p",
-      description: "Path to package.xml file to deploy"
+      char: 'p',
+      description: 'Path to package.xml file to deploy'
     }),
     filter: flags.boolean({
-      char: "f",
+      char: 'f',
       default: false,
-      description: "Filter metadatas before deploying"
+      description: 'Filter metadatas before deploying'
     }),
     destructivepackagexml: flags.string({
-      char: "k",
-      description: "Path to destructiveChanges.xml file to deploy"
+      char: 'k',
+      description: 'Path to destructiveChanges.xml file to deploy'
     }),
     testlevel: flags.enum({
-      char: "l",
-      default: "RunLocalTests",
+      char: 'l',
+      default: 'RunLocalTests',
       options: [
-        "NoTestRun",
-        "RunSpecifiedTests",
-        "RunLocalTests",
-        "RunAllTestsInOrg"
+        'NoTestRun',
+        'RunSpecifiedTests',
+        'RunLocalTests',
+        'RunAllTestsInOrg'
       ],
-      description: messages.getMessage("testLevel")
+      description: messages.getMessage('testLevel')
     }),
     debug: flags.boolean({
-      char: "d",
+      char: 'd',
       default: false,
-      description: messages.getMessage("debugMode")
+      description: messages.getMessage('debugMode')
     })
   };
 
@@ -77,9 +77,9 @@ export default class DxSources extends SfdxCommand {
     const packageXml = this.flags.packagexml || null;
     const filter = this.flags.filter || false ;
     const destructivePackageXml = this.flags.destructivepackagexml || null;
-    const testlevel = this.flags.testlevel || "RunLocalTests";
+    const testlevel = this.flags.testlevel || 'RunLocalTests';
     const debugMode = this.flags.debug || false;
-    this.configInfo = await getConfig("branch");
+    this.configInfo = await getConfig('branch');
 
     const essentialsRes = await checkSfdxPlugin('sfdx-essentials');
     this.ux.log(essentialsRes.message);
@@ -98,11 +98,11 @@ export default class DxSources extends SfdxCommand {
       destructivePackageXml ||
       process.env.PACKAGE_XML_TO_DELETE ||
       this.configInfo.packageXmlToDelete ||
-      fs.existsSync("./manifest/destructiveChanges.xml")
-        ? "./manifest/destructiveChanges.xml"
-        : fs.existsSync("./destructiveChanges.xml")
-        ? "./destructiveChanges.xml"
-        : "./config/destructiveChanges.xml";
+      fs.existsSync('./manifest/destructiveChanges.xml')
+        ? './manifest/destructiveChanges.xml'
+        : fs.existsSync('./destructiveChanges.xml')
+        ? './destructiveChanges.xml'
+        : './config/destructiveChanges.xml';
     if (fs.existsSync(packageDeletedXmlFile)) {
       // Create empty deployment file because of sfdx limitation
       // cf https://gist.github.com/benahm/b590ecf575ff3c42265425233a2d727e
@@ -113,71 +113,71 @@ export default class DxSources extends SfdxCommand {
       );
       const tmpDir = path.join(
         os.tmpdir(),
-        "sfdx-hardis-" + parseFloat(Math.random().toString())
+        'sfdx-hardis-' + parseFloat(Math.random().toString())
       );
       await fs.ensureDir(tmpDir);
-      const emptyPackageXmlFile = path.join(tmpDir, "package.xml");
+      const emptyPackageXmlFile = path.join(tmpDir, 'package.xml');
       await fs.writeFile(
         emptyPackageXmlFile,
         `<?xml version="1.0" encoding="UTF-8"?>
         <Package xmlns="http://soap.sforce.com/2006/04/metadata">
           <version>${CONSTANTS.API_VERSION}</version>
         </Package>`,
-        "utf8"
+        'utf8'
       );
       await fs.copy(
         packageDeletedXmlFile,
-        path.join(tmpDir, "destructiveChanges.xml")
+        path.join(tmpDir, 'destructiveChanges.xml')
       );
       const deployDelete =
         `sfdx force:mdapi:deploy -d ${tmpDir}` +
-        " --wait 60" +
-        " --testlevel NoTestRun" +
-        " --ignorewarnings" + // So it does not fail in case metadata is already deleted
-        (check ? " --checkonly" : "") +
-        (debugMode ? " --verbose" : "");
+        ' --wait 60' +
+        ' --testlevel NoTestRun' +
+        ' --ignorewarnings' + // So it does not fail in case metadata is already deleted
+        (check ? ' --checkonly' : '') +
+        (debugMode ? ' --verbose' : '');
       const deployDeleteRes = await execCommand(deployDelete, this, {
         output: true,
         debugMode,
         fail: true
       });
       await fs.remove(tmpDir);
-      let deleteMsg = "";
+      let deleteMsg = '';
       if (deployDeleteRes.status === 0) {
         destructiveProcessed = true;
         deleteMsg =
-          "[sfdx-hardis] Successfully deployed destructive changes to Salesforce org";
+          '[sfdx-hardis] Successfully deployed destructive changes to Salesforce org';
         this.ux.log(c.green(deleteMsg));
       } else {
         deleteMsg =
-          "[sfdx-hardis] Unable to deploy destructive changes to Salesforce org";
+          '[sfdx-hardis] Unable to deploy destructive changes to Salesforce org';
         this.ux.log(c.red(deployDeleteRes.errorMessage));
       }
     } else {
       uxLog(
         this,
-        "No destructivePackageXml found so no destructive deployment has been performed"
+        'No destructivePackageXml found so no destructive deployment has been performed'
       );
     }
 
     // Deploy sources
-    let packageXmlFile =
+    const packageXmlFile =
       packageXml ||
       process.env.PACKAGE_XML_TO_DEPLOY ||
       this.configInfo.packageXmlToDeploy ||
-      fs.existsSync("./manifest/package.xml")
-        ? "./manifest/package.xml"
-        : fs.existsSync("./package.xml")
-        ? "./package.xml"
-        : "./config/package.xml";
+      fs.existsSync('./manifest/package.xml')
+        ? './manifest/package.xml'
+        : fs.existsSync('./package.xml')
+        ? './package.xml'
+        : './config/package.xml';
     if (fs.existsSync(packageXmlFile)) {
-        let deployDir = '.'
+        let deployDir = '.';
         // Filter if necessary
         if (filter) {
             const tmpDir = path.join(os.tmpdir(), 'sfdx-hardis-deploy-') + Math.random().toString(36).slice(-5);
-            const filterCommand = 'sfdx essentials:metadata:filter-from-packagexml'+
+            const filterCommand = 'sfdx essentials:metadata:filter-from-packagexml' +
             ` -i ${deployDir}` +
-            ` -p ${packageXmlFile}`+
+            ` -p ${packageXmlFile}` +
             ` -o ${tmpDir}`;
             deployDir = tmpDir ;
             await execCommand(filterCommand, this, {
@@ -187,38 +187,38 @@ export default class DxSources extends SfdxCommand {
               });
         }
         // Perform deployment
-      const deployCommand =
-        `sfdx force:mdapi:deploy` +
-        ` --deploydir ${deployDir}`
-        " --wait 60" +
+        const deployCommand =
+        'sfdx force:mdapi:deploy' +
+        ` --deploydir ${deployDir}` +
+        ' --wait 60' +
         ` --testlevel ${testlevel}` +
-        (check ? " --checkonly" : "") +
-        (debugMode ? " --verbose" : "");
-      const deployRes = await execCommand(deployCommand, this, {
+        (check ? ' --checkonly' : '') +
+        (debugMode ? ' --verbose' : '');
+        const deployRes = await execCommand(deployCommand, this, {
         output: true,
         debugMode,
         fail: true
       });
-      let message = "";
-      if (deployRes.status === 0) {
+        let message = '';
+        if (deployRes.status === 0) {
         deployProcessed = true;
         message =
-          "[sfdx-hardis] Successfully deployed sfdx project sources to Salesforce org";
+          '[sfdx-hardis] Successfully deployed sfdx project sources to Salesforce org';
         this.ux.log(c.green(message));
       } else {
         message =
-          "[sfdx-hardis] Unable to deploy sfdx project sources to Salesforce org";
+          '[sfdx-hardis] Unable to deploy sfdx project sources to Salesforce org';
         this.ux.log(c.red(deployRes.errorMessage));
       }
     } else {
-      uxLog(this, "No package.xml found so no deployment has been performed");
+      uxLog(this, 'No package.xml found so no deployment has been performed');
     }
 
     return {
       orgId: this.org.getOrgId(),
       deployProcessed,
       destructiveProcessed,
-      outputString: ""
+      outputString: ''
     };
   }
 }